feat: Use commit message and PR title to determine story
In some use cases, PR titles contain the expected story string,
and they may be inside brackets. We should allow bracketized jira
stories as well as consider the PR title contents when evaluating
the message for the story. If both are present, we will still
prefer the commit message over PR title.